tests/wpt/tests/css/css-anchor-position/position-area-interpolation.html

<!DOCTYPE html>
<title>CSS Anchor Positioning: position-area interpolation</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-position/#position-area">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>
<body>
<script>
  test_no_interpolation({
    property: "position-area",
    from: "none",
    to: "span-all left"
  });
</script>
